+++ debug.log
@@ -6,3 +6,5 @@
 [1026/172025.682:ERROR:crash_report_database_win.cc(457)] unexpected header
 [1027/093924.436:ERROR:crash_report_database_win.cc(457)] unexpected header
 [1027/120258.090:ERROR:crash_report_database_win.cc(457)] unexpected header
+[1027/185430.753:ERROR:crash_report_database_win.cc(457)] unexpected header
+[1028/085040.463:ERROR:crash_report_database_win.cc(457)] unexpected header
