+++ src/components/App.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { useState, useEffect } from 'react';
 // import PropTypes from 'prop-types';
 
 import {
@@ -7,23 +6,11 @@ import {
   Switch,
   Route,
   Link
-} from "react-router-dom"; import Products from './products';
-import axios from 'axios';
+} from "react-router-dom";
+import Products from './products';
 import '../App.css';
 
 function App() {
-  const [data, setData] = useState({ products: [] });
-  useEffect(() => {
-    const fetchData = async () => {
-      const result = await axios(
-        'https://pokeapi.co/api/v2/pokemon?limit=151',
-      );
-      console.log(result.data.results)
-      setData(result.data);
-    };
-
-    fetchData();
-  }, []);
 
   return (
     <Router>
@@ -38,13 +25,7 @@ function App() {
             </li>
           </ul>
         </nav>
-        {/* <ul>
-          {data.results.map(item => (
-            <li key={item}>
-              <a href={item}>{item}</a>
-            </li>
-          ))}
-        </ul> */}
+        {/* {console.log(data.results)} */}
 
         {/* A <Switch> looks through its children <Route>s and
           renders the first one that matches the current URL. */}
