+++ src/components/Products.js
@@ -1,8 +1,24 @@
 import React from 'react';
 import { connect } from 'react-redux';
+import axios from 'axios';
+import { useState, useEffect } from 'react';
+
+
 
 const Products = ({ products }) => {
-  console.log(products)
+  const [data, setData] = useState({ products });
+  useEffect(() => {
+    const fetchData = async () => {
+      const result = await axios(
+        'https://pokeapi.co/api/v2/pokemon?limit=151',
+      );
+      console.log(products.products)
+
+      setData(result);
+    };
+
+    fetchData();
+  }, []);
 
   // const handleFilterChange = e => {
   //   const filter = e.target.value;
@@ -11,7 +27,7 @@ const Products = ({ products }) => {
 
   return (
     <div>
-      <h1>Home...</h1>
+      <h1>{data.results}</h1>
     </div>
   )
 }
