+++ src/index.js
@@ -7,9 +7,7 @@ import App from './components/App';
 import reportWebVitals from './reportWebVitals';
 import rootReducer from './reducers/index';
 
-
 const store = createStore(rootReducer);
-
 ReactDOM.render(
   <Provider store={store}>
     <App />
