+++ src/reducers/products.js
@@ -1,24 +1,7 @@
-import axios from 'axios';
-
-function getData() {
-  axios
-    .all([
-      axios.get('https://pokeapi.co/api/v2/ability/?limit=20&offset=20')
-    ])
-    .then(axios.spread((products) => showOutput(products)))
-    .catch(err => console.error(err));
-}
-
-function showOutput(res) {
-  return res
-}
-// const data = getData();
-// console.log(data)
-
 const initialState = {
-  // filter: 'All',
-  products: getData()
-};
+  filter: 'All',
+  products: [],
+}
 
 function products(state = initialState, action) {
   switch (action.type) {
@@ -33,18 +16,3 @@ function products(state = initialState, action) {
 }
 
 export default products;
-
-
-
-
-
-
-
-
-// getData()
-
-
-
-// function products(state = [], action) {
-//   console.log('new')
-// }
